module7-task1 | homework | FINAL UPDATE + BUGFIX
